store/requirements.txt
Django==5.1.6
django-cockroachdb==5.1
psycopg2-binary==2.9.10
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
tonsdk==1.0.15
